Moved ShapeNet to kaolin.io
