+++ kaolin/io/__init__.py
@@ -1,4 +1,5 @@
 from . import dataset
 from . import materials
 from . import obj
+from . import shapenet
 from . import usd
